Use greater than 0 as the comparison.
This is a question of readability, context and numerical comparison...
An empty list has 0 elements, thus a non-empty list has non-0 elements; a non-0 positive value.

Logically for the presence of something isn't that there are no items; length != 0; but rather
that there are items and as such, length of the array is greater than 0;
The context of comparison is the count of items has to be a positive value.

For readability, where possible, making a positive statement is easier to read and write than making a negative
statement to determine a truth.